Refactor redux action for Charity item
