+++ src/components/container/CharityList/index.js
@@ -2,11 +2,9 @@ import React, { useEffect } from 'react';
 import { compose } from 'recompose';
 import styled from 'styled-components';
 import { connect } from 'react-redux';
-import {
-	getCharityList,
-	getPaymentList,
-	payDonation
-} from '~/state/ducks/actions/charityPayment';
+import { useTransition, animated } from 'react-spring';
+
+import { getCharityList } from '~/state/ducks/actions/charityPayment';
 import { message, Spin } from 'antd';
 import { donationDataAction } from '~/state/ducks/donationData';
 import { summaryDonations } from '~/helpers';
@@ -17,8 +15,9 @@ const Wrapper = styled.div`
 	margin-top: 30px;
 `;
 
-const Item = styled.div`
+const Item = styled(animated.div)`
 	margin-bottom: 50px;
+	will-change: opacity, transform;
 `;
 
 const LoadingWrapper = styled.div`
@@ -27,11 +26,22 @@ const LoadingWrapper = styled.div`
 `;
 
 const CharityList = props => {
-	const { getCharityList, getPaymentList, updateDonation } = props;
+	const {
+		getCharityList,
+		getPaymentList,
+		updateDonation,
+	} = props;
 
 	const { reducerData, loadList, loadSuccess, loadFail } = useMiniReducer();
 
 	const { data, loading } = reducerData;
+
+	const transitions = useTransition(data, item => item.id, {
+		from: { o: 0, y: 20 },
+		enter: { o: 1, y: 0 },
+		leave: { o: 0, y: 0 },
+		trail: 200
+	});
 
 	// fetch charity list and payment
 	useEffect(() => {
@@ -63,14 +73,23 @@ const CharityList = props => {
 
 	return (
 		<Wrapper className="row">
-			{data.map(charity => (
-				<Item className="col-12 col-sm-6" key={charity.id}>
-					<CharityItem charity={charity} />
+			{transitions.map(({ item, key, props: aniProps }) => (
+				<Item
+					className="col-12 col-sm-6"
+					key={key}
+					style={{
+						opacity: aniProps.o,
+						transform: aniProps.y.interpolate(y => `translate3d(0, ${y}px, 0)`)
+					}}
+				>
+					<CharityItem charity={item} />
 				</Item>
 			))}
 		</Wrapper>
 	);
 };
+
+const { updateDonation, getPaymentList } = donationDataAction;
 
 export default compose(
 	connect(
@@ -78,8 +97,7 @@ export default compose(
 		{
 			getCharityList,
 			getPaymentList,
-			payDonation,
-			updateDonation: donationDataAction.updateDonation
+			updateDonation
 		}
 	)
 )(CharityList);
