+++ src/components/presentation/CharityList/index.js
@@ -1,0 +1,2 @@
+export { default as CharityTotal } from './CharityTotal';
+export { default as OverlayUnit } from './OverlayUnit';
