+++ src/state/ducks/actions/charityPayment/types.js
@@ -1,4 +1,2 @@
 export const LOAD_CHARITY_LIST = 'charityPayment/LOAD_CHARITY_LIST';
-export const LOAD_PAYMENT_LIST = 'charityPayment/LOAD_PAYMENT_LIST';
 
-export const PAY_DONATION = 'charityPayment/LOAD_PAYMENT_LIST';
