+++ src/state/ducks/donationData/reducers.js
@@ -1,22 +1,26 @@
-import * as types from "./types";
-import produce from "immer";
+import * as types from './types';
+import produce from 'immer';
 
 const initialState = {
-  donate: 0,
-  message: ""
+	donate: 0,
+	message: '',
+	payments: []
 };
 
 const reducer = produce((draft, { payload, type }) => {
-  switch (type) {
-    case types.UPDATE_TOTAL_DONATE:
-      draft.donate = draft.donate + payload.amount;
-      return;
-    case types.UPDATE_MESSAGE:
-      draft.message = payload.message;
-    break;
-    default:
-      return draft;
-  }
+	switch (type) {
+		case types.UPDATE_TOTAL_DONATE:
+			draft.donate = draft.donate + payload.amount;
+			return;
+		case types.LOAD_PAYMENT_LIST_SUCCESS:
+			draft.payments = payload.payments;
+			return;
+		case types.UPDATE_MESSAGE:
+			draft.message = payload.message;
+			break;
+		default:
+			return draft;
+	}
 }, initialState);
 
 export default reducer;
