+++ src/state/ducks/donationData/types.js
@@ -1,2 +1,7 @@
 export const UPDATE_TOTAL_DONATE = 'donationData/UPDATE_TOTAL_DONATE';
+export const UPDATE_PAYMENTS = 'donationData/UPDATE_PAYMENTS';
 export const UPDATE_MESSAGE = 'donationData/UPDATE_MESSAGE';
+
+export const LOAD_PAYMENT_LIST = 'donationData/LOAD_PAYMENT_LIST';
+export const LOAD_PAYMENT_LIST_SUCCESS = 'donationData/LOAD_PAYMENT_LIST_SUCCESS';
+export const PAY_DONATION = 'donationData/PAY_DONATION';
